Rename to lexer
