+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod ast;
-mod tokenize;
+mod lexer;
 
 pub fn parse(_input: &str) -> ast::Stylesheet {
     todo!()
